+++ src/main/java/com/example/banking_system/domain/loan/dto/UpdateLoanPolicyRequest.java
@@ -7,6 +7,7 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -19,7 +20,7 @@ public class UpdateLoanPolicyRequest {
     private Integer durationMonths;
 
     @DecimalMin(value = "0.0", inclusive = false, message = "interest rate must be greater than 0")
-    private Double interestRate;
+    private BigDecimal interestRate;
 
     private LoanType loanType;
